refactor: move user related api calls from user-api-service to shared auth-lib-service
